Chessboard [Material and no material + React-srping]
Package:
+ "@react-spring/three

I wanted to load in my own textures and then use a material on the model. Doing so turned out that the texture that were loaded by the useTexture method of 'drei' were flipped on the Y-axis. This strange behaviour let to some confusion. I added both types of chessboards to see which one I will use later on in the development